Renaming sci spin box file.
